refactor(docs): checkbox dx
