+++ apps/docs/content/components/checkbox/controlled.ts
@@ -1,19 +1,4 @@
-const App = `import {Checkbox} from "@nextui-org/react";
-
-export default function App() {
-  const [isSelected, setIsSelected] = React.useState(false);
-
-  return (
-    <div className="flex flex-col gap-2">
-      <Checkbox isSelected={isSelected} onValueChange={setIsSelected}>
-        Subscribe (controlled)
-      </Checkbox>
-      <p className="text-default-500">
-        Selected: {isSelected ? "true" : "false"}
-      </p>
-    </div>
-  );
-}`;
+import App from "./controlled.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
