+++ apps/docs/content/components/checkbox/disabled.raw.jsx
@@ -1,0 +1,12 @@
+import {Checkbox} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex gap-4">
+      <Checkbox isDisabled>Option</Checkbox>
+      <Checkbox defaultSelected isDisabled>
+        Option
+      </Checkbox>
+    </div>
+  );
+}
