+++ apps/docs/content/components/checkbox/indeterminate.raw.jsx
@@ -1,0 +1,5 @@
+import {Checkbox} from "@nextui-org/react";
+
+export default function App() {
+  return <Checkbox isIndeterminate>Option</Checkbox>;
+}
