+++ apps/docs/content/components/checkbox/indeterminate.ts
@@ -1,10 +1,4 @@
-const App = `import {Checkbox} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Checkbox isIndeterminate>Option</Checkbox>
-  );
-}`;
+import App from "./indeterminate.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
