+++ apps/docs/content/components/checkbox/line-through.raw.jsx
@@ -1,0 +1,9 @@
+import {Checkbox} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Checkbox defaultSelected lineThrough>
+      Option
+    </Checkbox>
+  );
+}
