+++ apps/docs/content/components/checkbox/line-through.ts
@@ -1,10 +1,4 @@
-const App = `import {Checkbox} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Checkbox defaultSelected lineThrough>Option</Checkbox>
-  );
-}`;
+import App from "./line-through.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
