+++ apps/docs/content/components/checkbox/radius.raw.jsx
@@ -1,0 +1,23 @@
+import {Checkbox} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex gap-4">
+      <Checkbox defaultSelected radius="full">
+        Full
+      </Checkbox>
+      <Checkbox defaultSelected radius="lg">
+        Large
+      </Checkbox>
+      <Checkbox defaultSelected radius="md">
+        Medium
+      </Checkbox>
+      <Checkbox defaultSelected radius="sm">
+        Small
+      </Checkbox>
+      <Checkbox defaultSelected radius="none">
+        None
+      </Checkbox>
+    </div>
+  );
+}
